Fix: FAILED: updater_ramdisk.img

+++ ohos/images/mkimage/mkcpioimage.py
@@ -170,7 +170,11 @@ def parse_resource_config(resource_config_file_path):
             if each_section[0] == DTC_510:
                 dtc_510_source_path = source_path
             if os.path.exists(source_path):
-                shutil.copy(source_path, target_path)
+                if os.path.exists(target_path):
+                    if not filecmp.cmp(source_path, target_path):
+                        shutil.copy(source_path, target_path)
+                else:
+                    shutil.copy(source_path, target_path)
             else:
                 if each_section[0] not in [DTC_419, DTC_510]:
                     print("Error: source file does not exist! path: %s" %
